day 07 part1 (Camel Cards - poor mans poker (without suites))
Part 1 is quiet simple. The tricky bit is the comparison between Strengths and Hands.

Part 2 introduces Joker, which can replace any other card, but not fully, although it can "fill up" a hand to a stronger one, the card itself will count as the weakest when comparing hands with the same type.